+++ sdk/core/azure-core-amqp/src/test/java/com/azure/core/amqp/implementation/handler/ConnectionHandlerTest.java
@@ -11,6 +11,7 @@ import com.azure.core.amqp.implementation.ClientConstants;
 import com.azure.core.amqp.implementation.ConnectionOptions;
 import com.azure.core.credential.TokenCredential;
 import com.azure.core.util.ClientOptions;
+import com.azure.core.util.Header;
 import com.azure.core.util.UserAgentUtil;
 import org.apache.qpid.proton.Proton;
 import org.apache.qpid.proton.amqp.Symbol;
@@ -31,14 +32,18 @@ import org.mockito.MockitoAnnotations;
 import reactor.core.scheduler.Scheduler;
 import reactor.test.StepVerifier;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import static com.azure.core.amqp.implementation.handler.ConnectionHandler.AMQPS_PORT;
 import static com.azure.core.amqp.implementation.handler.ConnectionHandler.FRAMEWORK;
 import static com.azure.core.amqp.implementation.handler.ConnectionHandler.MAX_FRAME_SIZE;
 import static com.azure.core.amqp.implementation.handler.ConnectionHandler.PLATFORM;
+import static com.azure.core.amqp.implementation.handler.ConnectionHandler.PRODUCT;
 import static com.azure.core.amqp.implementation.handler.ConnectionHandler.USER_AGENT;
+import static com.azure.core.amqp.implementation.handler.ConnectionHandler.VERSION;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -50,11 +55,16 @@ import static org.mockito.Mockito.when;
 
 public class ConnectionHandlerTest {
     private static final String APPLICATION_ID = "some-random-app-id";
-    private static final ClientOptions CLIENT_OPTIONS = new ClientOptions().setApplicationId(APPLICATION_ID);
     private static final String CONNECTION_ID = "some-random-id";
     private static final String HOSTNAME = "hostname-random";
-    private static final String PRODUCT = "test";
+    private static final String CLIENT_PRODUCT = "test";
     private static final String CLIENT_VERSION = "1.0.0-test";
+    private static final List<Header> HEADER_LIST = Collections.singletonList(
+        new Header("foo-bar", "some-values"));
+    private static final ClientOptions CLIENT_OPTIONS = new ClientOptions()
+        .setApplicationId(APPLICATION_ID)
+        .setHeaders(HEADER_LIST);
+
     private static final SslDomain.VerifyMode VERIFY_MODE = SslDomain.VerifyMode.VERIFY_PEER_NAME;
 
     private final SslPeerDetails peerDetails = Proton.sslPeerDetails(HOSTNAME, 2919);
@@ -74,8 +84,8 @@ public class ConnectionHandlerTest {
 
         this.connectionOptions = new ConnectionOptions(HOSTNAME, tokenCredential,
             CbsAuthorizationType.SHARED_ACCESS_SIGNATURE, AmqpTransportType.AMQP, new AmqpRetryOptions(),
-            ProxyOptions.SYSTEM_DEFAULTS, scheduler, CLIENT_OPTIONS, VERIFY_MODE);
-        this.handler = new ConnectionHandler(CONNECTION_ID, PRODUCT, CLIENT_VERSION, connectionOptions, peerDetails);
+            ProxyOptions.SYSTEM_DEFAULTS, scheduler, CLIENT_OPTIONS, VERIFY_MODE, CLIENT_PRODUCT, CLIENT_VERSION);
+        this.handler = new ConnectionHandler(CONNECTION_ID, connectionOptions, peerDetails);
     }
 
     @AfterEach
@@ -90,29 +100,23 @@ public class ConnectionHandlerTest {
 
     @Test
     void constructorNull() {
-        assertThrows(NullPointerException.class, () -> new ConnectionHandler(null, PRODUCT,
-            CLIENT_VERSION, connectionOptions, peerDetails));
-        assertThrows(NullPointerException.class, () -> new ConnectionHandler(CONNECTION_ID, null,
-            CLIENT_VERSION, connectionOptions, peerDetails));
-        assertThrows(NullPointerException.class, () -> new ConnectionHandler(CONNECTION_ID, PRODUCT,
-            null, connectionOptions, peerDetails));
-        assertThrows(NullPointerException.class, () -> new ConnectionHandler(CONNECTION_ID, PRODUCT,
-            CLIENT_VERSION, null, peerDetails));
-        assertThrows(NullPointerException.class, () -> new ConnectionHandler(CONNECTION_ID, PRODUCT,
-            CLIENT_VERSION, connectionOptions, null));
+        assertThrows(NullPointerException.class, () -> new ConnectionHandler(null, connectionOptions, peerDetails));
+        assertThrows(NullPointerException.class, () -> new ConnectionHandler(CONNECTION_ID, null, peerDetails));
+        assertThrows(NullPointerException.class, () -> new ConnectionHandler(CONNECTION_ID, connectionOptions, null));
     }
 
     @Test
     void applicationIdNotSet() {
         // Arrange
-        final ClientOptions clientOptions = new ClientOptions();
-        final String expected = UserAgentUtil.toUserAgentString(null, PRODUCT, CLIENT_VERSION, null);
+        final ClientOptions clientOptions = new ClientOptions()
+            .setHeaders(HEADER_LIST);
+        final String expected = UserAgentUtil.toUserAgentString(null, CLIENT_PRODUCT, CLIENT_VERSION, null);
         final ConnectionOptions options = new ConnectionOptions(HOSTNAME, tokenCredential,
             CbsAuthorizationType.SHARED_ACCESS_SIGNATURE, AmqpTransportType.AMQP, new AmqpRetryOptions(),
-            ProxyOptions.SYSTEM_DEFAULTS, scheduler, clientOptions, VERIFY_MODE);
+            ProxyOptions.SYSTEM_DEFAULTS, scheduler, clientOptions, VERIFY_MODE, CLIENT_PRODUCT, CLIENT_VERSION);
 
         // Act
-        final ConnectionHandler handler = new ConnectionHandler(CONNECTION_ID, PRODUCT, CLIENT_VERSION, options, peerDetails);
+        final ConnectionHandler handler = new ConnectionHandler(CONNECTION_ID, options, peerDetails);
 
         // Assert
         final String userAgent = (String) handler.getConnectionProperties().get(USER_AGENT.toString());
@@ -122,24 +126,28 @@ public class ConnectionHandlerTest {
     @Test
     void applicationIdSet() {
         // Arrange
-        final String expected = UserAgentUtil.toUserAgentString(CLIENT_OPTIONS.getApplicationId(), PRODUCT,
+        final String expected = UserAgentUtil.toUserAgentString(CLIENT_OPTIONS.getApplicationId(), CLIENT_PRODUCT,
             CLIENT_VERSION, null);
 
         // Act
-        final ConnectionHandler handler = new ConnectionHandler(CONNECTION_ID, PRODUCT, CLIENT_VERSION,
-            connectionOptions, peerDetails);
+        final ConnectionHandler handler = new ConnectionHandler(CONNECTION_ID, connectionOptions, peerDetails);
 
         // Assert
         final String userAgent = (String) handler.getConnectionProperties().get(USER_AGENT.toString());
         assertEquals(expected, userAgent);
     }
 
+    /**
+     * Verifies that our connection properties are all set.
+     */
     @Test
     void createHandler() {
         // Arrange
         final Map<String, String> expected = new HashMap<>();
         expected.put(PLATFORM.toString(), ClientConstants.PLATFORM_INFO);
         expected.put(FRAMEWORK.toString(), ClientConstants.FRAMEWORK_INFO);
+        expected.put(PRODUCT.toString(), CLIENT_PRODUCT);
+        expected.put(VERSION.toString(), CLIENT_VERSION);
 
         // Assert
         assertEquals(HOSTNAME, handler.getHostname());