+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/EntityConditionality.java
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.ai.textanalytics.models;
+
+import com.azure.core.annotation.Immutable;
+
+/**
+ * The {@link EntityConditionality} model.
+ */
+@Immutable
+public enum EntityConditionality {
+    /** Enum value Hypothetical. */
+    HYPOTHETICAL,
+
+    /** Enum value Conditional. */
+    CONDITIONAL
+}