+++ sdk/spring/azure-spring-boot/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 3.3.0-beta.1 (Unreleased)
+### Key Bug Fixes
+- Fix bug of using closed `MessageProducer` and `MessageConsumer` when a link is force detached in azure-spring-boot-starter-servicebus-jms.
+
 ## 3.2.0 (2021-03-03)
 ### Breaking Changes
 - Remove `azure.activedirectory.b2c.oidc-enabled` property.